Shebang, no input and inspecting JSON objects
This commit fixes and the Shebang, moves it above the header (was below wrongly)
It also introduces JSON object inspection, usable with the first argument "inspect". This will
display object keys, array lengths or the values and their types.
Furthermore now no input is also possible, previously this broke the program.